合并了修改版本号8962, 9003 从 branches/CodeTest: 合并了修改版本号 从 branches/Code: 创意仓库二期-数据中心：创意仓库-我的活动，查看活动数据，订单数没有显示，当前已有订单（#2070）
........
合并了修改版本号9002 从 branches/Code:
创意仓库-我的活动，查看活动数据，当前活动商品销售有数据，点击查看销售清单无数据

........

........


git-svn-id: https://10.133.192.229/svn/CustomerManage/trunk/Code@9024 67192a19-bcbd-5648-baee-99b7ca3a6e28

+++ CPOS.DTO/Module/CreativityWarehouse/MarketingData/Response/GeEventItemListRD.cs
@@ -24,11 +24,11 @@ namespace JIT.CPOS.DTO.Module.CreativityWarehouse.MarketingData.Response
         public string SkuName { get; set; }
         public decimal price { get; set; }
         public decimal SalesPrice { get; set; }
-           public int Qty { get; set; }
-    public int KeepQty { get; set; }
-          public int SoldQty { get; set; }
-       public int InverTory { get; set; }
-        public int TotalSales { get; set; }
+        public int Qty { get; set; }
+        public int KeepQty { get; set; }
+        public int SoldQty { get; set; }
+        public int InverTory { get; set; }
+        public decimal TotalSales { get; set; }
         public int TurnoverRate { get; set; }
